+++ pkgs/matcher/test/pretty_print_unminified_test.dart
@@ -13,7 +13,7 @@ import 'dart:collection';
 
 import 'package:matcher/matcher.dart';
 import 'package:matcher/src/pretty_print.dart';
-import 'package:unittest/unittest.dart' show test, group;
+import 'package:unittest/unittest.dart' show test, group, expect;
 
 class DefaultToString {}